modificando pages
