+++ src/Components/Forms/CreateEvent/StepThreeFinal.jsx
@@ -1,0 +1,55 @@
+import React from 'react'
+
+class StepThree extends React.Component {
+  constructor () {
+    super()
+    this.state = { 
+      password: '', 
+      passwordConfirm: '' 
+    }
+    this.handlePasswordChanged = this.handlePasswordChanged.bind(this);
+    this.handlePasswordConfirmChanged = this.handlePasswordConfirmChanged.bind(this);
+  }
+
+  handlePasswordChanged (event) {
+    this.setState({password: event.target.value})
+  }
+
+  handlePasswordConfirmChanged (event) {
+    this.setState({passwordConfirm: event.target.value})
+  }
+
+  render () {
+    return (
+      <div>
+        <div className='row'>
+          <div className='six columns'>
+            <label>Password</label>
+            <input
+              className='u-full-width required'
+              placeholder='Password'
+              type='password'
+              onChange={this.handlePasswordChanged}
+              value={this.state.password}
+              autoFocus
+            />
+          </div>
+        </div>
+        <div className='row'>
+          <div className='six columns'>
+            <label>Confirm password</label>
+            <input
+              className='u-full-width'
+              placeholder='Confirm Password'
+              type='password'
+              onChange={this.handlePasswordConfirmChanged}
+              value={this.state.passwordConfirm}
+            />
+          </div>
+        </div>
+      </div>
+    )
+  }
+}
+
+export default StepThree;
