+++ src/Components/General/bannerTop.jsx
@@ -62,26 +62,26 @@ class BannerTop extends Component{
           <div class="collapse navbar-collapse" id="navbarNavDropdown">
             <ul class="nav navbar-nav">
               <li class="nav-item active">
-                <a class="nav-link" href="/">INICIO  |<span class="sr-only">(current)</span></a>
+                <a class="nav-link" href="/">Inicio  |<span class="sr-only">(current)</span></a>
               </li>
               
               <li class="nav-item">
-                <a class="nav-link" href="/events">EVENTOS  </a>
+                <a class="nav-link" href="/events">Eventos  </a>
               </li>
               
               <li class="nav-item">
-                <a class="nav-link" href="/announcements">CONVOCATORIA | </a>
+                <a class="nav-link" href="/announcements">Convocatoria | </a>
               </li>
               
               <li class="nav-item" class="nav dropdown">
-                <a class="nav-link dropdown-toggle" href="#" data-toggle="dropdown" role="button"  aria-haspopup="true" aria-expanded="false">OPCIONES... </a>
+                <a class="nav-link dropdown-toggle" href="#" data-toggle="dropdown" role="button"  aria-haspopup="true" aria-expanded="false">Opciones... </a>
                 <ul class="dropdown-menu">
                   
-                  <li><a class="nav-link" href="#">MIS INSCRIPCIONES</a></li>
-                  <li><a class="nav-link" href="#">MIS PROPUESTAS</a></li>
-                  <li><a class="nav-link" href="/Organizador_ActiveEvents">ORGANIZADOR</a></li>
-                  <li><a class="nav-link" href="/PresidenteEventos">PRESIDENTE</a></li>
-                  <li><a class="nav-link" href="/EvaluadorEventos">EVALUADOR</a></li>
+                  <li><a class="nav-link" href="#">Mis inscripciones</a></li>
+                  <li><a class="nav-link" href="#">Mis propuestas</a></li>
+                  <li><a class="nav-link" href="/organActiveEvents">Organizador</a></li>
+                  <li><a class="nav-link" href="/presidentEvents">Presidente</a></li>
+                  <li><a class="nav-link" href="/EvaluadorEventos">Evaluador</a></li>
                 </ul>
               </li>
             </ul>
