+++ src/Components/Jtable/Organizador_ActiveEventsTable.jsx
@@ -1,56 +1,71 @@
-import React, { Component } from 'react'
-import './../../styles/Jtab.css'
-class Organizador_ActiveEventsTable extends Component {
+import React, { Component } from 'react';
+import 'bootstrap/dist/css/bootstrap.css';
+
+
+class Organizador_ActiveEventsTable  extends Component {
    constructor(props) {
-    
       super(props) //since we are extending class Table so we have to use super in order to override Component class constructor
       console.log("rzwetxrytcvygbuhnj"+this.props);
-      
    }
    handleClick = () => {
     console.log('this is:', this);
   }
   
-   renderTableData() {
+   tableData() {
         return this.state.chupetinesGA.map((element, index) => {
-        const { listaEventos} = element //destructuring
+        const { listaEventos, propRec, propEval, programa} = element //destructuring
         return (
-            <tr >
+            <tr>
                 <td>{listaEventos}</td>
                 <td>
-                   <button class="btn_plus" onClick={this.handleClick} ><i class="fa fa-download"></i></button>
+                   <h1>
+                      <button class="btn_plus" onClick={this.handleClick} ><i class="fa fa-check-circle"></i></button>
+                    - <button class="btn_plus" onClick={this.handleClick} ><i class="fa fa-check-circle"></i></button>
+                    - <button class="btn_plus" onClick={this.handleClick} ><i class="fa fa-check-circle"></i></button>
+                  </h1>
                </td> 
-                </tr>
+               <td>
+                   <button class="btn_plus" onClick={this.handleClick} ><i class="fa fa-edit"></i></button>
+               </td> 
+               <td>
+                   <button class="btn_plus" onClick={this.handleClick} ><i class="fa fa-plus"></i></button>
+               </td> 
+               
+               <td>
+                   <button class="btn_plus" onClick={this.handleClick} ><i class="fa fa-play"></i></button>
+               </td> 
+               <td>
+                   <button class="btn_plus" onClick={this.handleClick} ><i class="fa fa-times-circle"></i></button>
+               </td> 
+                
+            </tr>
         )
         })
     }
-    renderTableHeader() {
-      return (
-         <tr>
-             <th width="70%">Lista de eventos</th>
-             <th width="30%">Reporte </th>
-             
-         </tr>
-
-     )
-     }
   
      render() {
         this.state = this.props.data
         return (
-            
-           <div>
-              <br/>
-              <h1 id='title'>Lista de eventos históricos</h1><br></br>
-              <table id='chupetinesGA'>
-                 <tbody>
-                    {this.renderTableHeader()}
-                    {this.renderTableData()}
-                 </tbody>
+           <div >
+              <h2 class="card-title" id='title' >Lista de Eventos activos</h2>
+              <div  class="table-responsive">
+              <table class="table-light" id='chupetinesGA'>
+                  <thead class="thead-light">
+                  <tr>
+                     <th scope="col">Lista de eventos</th>
+                     <th scope="col">Prop. rec. -> En eval. -> Prog. comp. </th>
+                     <th scope="col">Editar</th>
+                     <th scope="col">Seg. de fases</th>
+                     <th scope="col">Publicar evento</th>
+                     <th scope="col">Cancelar</th>
+                  </tr>
+               </thead>
+              <tbody>{this.tableData()}</tbody>
               </table>
+              </div>
            </div>
         )
      }
 }
 
-export default Organizador_ActiveEventsTable //exporting a component make it reusable and this is the beauty of react
+export default Organizador_ActiveEventsTable  //exporting a component make it reusable and this is the beauty of react
