+++ src/Pages/AnnouncementPage.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import BannerTop from '../Components/General/bannerTop';
 import BannerBottom from '../Components/General/bannerBottom';
-import Organizador_ActiveEventsTable from '../Components/Jtable/Organizador_ActiveEventsTable';
 class AnnouncementPage extends Component{
     state = {
         bannTop : BannerTop,
