+++ src/Pages/Login.js
@@ -2,9 +2,7 @@ import React, { Component } from 'react';
 import BannerLogin from '../Components/General/bannerLogin';
 import BannerBottom from '../Components/General/bannerBottom'
 import '../styles/style_record.css'; 
-import 'react-tabs/style/react-tabs.css';
-import 'react-table/react-table.css';
-import './../styles/style_gig_tittle.css';
+
 
 class Login extends Component{
     state = {
@@ -14,28 +12,45 @@ class Login extends Component{
       }
    
     render(){
-        return(<div> 
-            <body>
+        return(<div>            
+            <body>  
+                <div class="component-header"  width="300">
+                <a class="component-logo customizable chart" href='/' title="Volver a pagina principal">
+                    <img class="component-logo" src="logo.png"  width="240"/> 
+                </a>
+                </div>                
                 <section class="container-fluid bg">
                     <section class="row justify-content-center">
                         <section class="col-12  col-sm-6 col-md-3">
-                        <form class="form-container">
+                        <div class="form-container">
+                        <form action="/organActiveEvents" >
                             <div class="form-group">
-                                <label for="exampleInputEmail1">Email address</label>
-                                <input type="email" class="form-control" id="exampleInputEmail1" aria-describedby="emailHelp" placeholder="Enter email"/>
-                                <small id="emailHelp" class="form-text text-muted">We'll never share your email with anyone else.</small>
+                                <label for="exampleInputEmail1">Usuario</label>
+                                <input type="text" class="form-control" id="exampleInputEmail1"  placeholder="Ingresar usuario"/>
                             </div>
                             <div class="form-group">
-                                <label for="exampleInputPassword1">Password</label>
-                                <input type="password" class="form-control" id="exampleInputPassword1" placeholder="Password"/>
+                                <label for="exampleInputPassword1">Contraseña</label>
+                                <input type="password" class="form-control" id="exampleInputPassword1" placeholder="Contraseña"/>
                             </div>
-                            <button type="submit" class="btn btn-primary btn-block">Submit</button>
+                            <button type="submit" class="btn btn-primary btn-block">Iniciar sesion</button>
                             
                         </form>
+                        <div class="right-box">
+                            <br/>
+                            <label for="exampleInputEmail1"  style={{textAlignVertical: "center",textAlign: "center"}}>Ingresar con cuenta gmail</label>
+                            <section>
+                            <script src= "./login.js"></script>
+                                <modal width="100" height="300" effect="fadeInUp" onClickAway={() => this.closeModal()}>                                    
+                                    <div class="g-signin2" data-onsuccess="onSignIn" ></div>
+                                </modal>
+                            </section>
+                        </div>
+                        </div>
                         </section>
                     </section>
                 </section>
             </body>
+            <this.state.bannBot />
         </div>)
     }
 }
