+++ src/styles/style_sheets.css
@@ -1,17 +1,33 @@
 .mybutton{
-    border-radius: 4px;
-    float: left;
+    border-radius: 6px;
     border: none;
     outline:none;
     cursor: pointer;
     padding: 10x 15px;
     font-family: sans-serif;
-    font-size: 20px;
+    font-size: 15px;    
     background-color: #002D3D;
     color: #FFFFFF;
-    text-align: center;            
+    text-align: center; 
+    width: 110px;
+    height:28px;    
+    line-height: 14px;       
 }
-
+.btnAdd{
+    border-radius: 5px;
+    border: none;
+    outline:none;
+    cursor: pointer;
+    padding: 10x 15px;
+    font-family: sans-serif;
+    font-size: 13px;    
+    background-color: #002D3D;
+    color: #FFFFFF;
+    text-align: center; 
+    width: 66px;
+    height:23px;    
+    line-height: 10px; 
+}
 .mybutton:hover {
     background-color: #E6E9EE;
     color: #002D3D;
@@ -20,11 +36,14 @@
 .form-container{
     position: absolute; 
     top: 20vh;
-    background: #fff;
-    padding: 30px;
+    background: rgba(255,255,255,.8);
+    width: 300px;
+    padding: 40px;
     border-radius: 10px;
     box-shadow: 0px 0px 10px 0px #000;
+    
 }
+
 
 .bg{
     background: url('../Components/Images/bg.jpg') no-repeat;
@@ -36,4 +55,23 @@
     .bg{
         background-size: 300%;
     }
+}
+
+.component-header{
+    right: 0;
+    left: 0;
+    top: 0;
+    position:absolute;
+    height: 62px;
+    line-height: 62px;
+    display:flex;
+    margin-right: auto;
+    margin-left: auto;
+}
+
+
+.component-header .component-logo.chart{
+    height: 50px;    
+    left: 5px;
+    top: 6px;
 }
